+++ groups/bdl/bdlt/bdlt_monthofyear.t.cpp
@@ -360,7 +360,7 @@ if (veryVerbose)
             ASSERT(out);
 
             const char *const OD  = out.data();
-            const int         LOD = out.length();
+            const int         LOD = static_cast<int>(out.length());
 
             In in(OD, LOD);
             ASSERT(in);
@@ -402,7 +402,7 @@ if (veryVerbose)
                 LOOP_ASSERT(i, &out == &rvOut);
                 LOOP_ASSERT(i, out);
                 const char *const OD  = out.data();
-                const int         LOD = out.length();
+                const int         LOD = static_cast<int>(out.length());
 
                 // Verify that each new value overwrites every old value and
                 // that the input stream is emptied, but remains valid.
@@ -438,7 +438,7 @@ if (veryVerbose)
         {
             Out               out(VERSION_SELECTOR, &allocator);
             const char *const OD  = out.data();
-            const int         LOD = out.length();
+            const int         LOD = static_cast<int>(out.length());
             ASSERT(0 == LOD);
 
             for (int i = 0; i < NUM_VALUES; ++i) {
@@ -485,7 +485,7 @@ if (veryVerbose)
             ASSERT(out);
 
             const char *const OD  = out.data();
-            const int         LOD = out.length();
+            const int         LOD = static_cast<int>(out.length());
             ASSERT(0 < LOD);
 
             for (int i = 0; i < NUM_VALUES; ++i) {
@@ -535,17 +535,17 @@ if (veryVerbose)
             Out& rvOut1 = bdexStreamOut(out, X1, VERSION);
             ASSERT(&out == &rvOut1);
             ASSERT(out);
-            const int         LOD1 = out.length();
+            const int         LOD1 = static_cast<int>(out.length());
 
             Out& rvOut2 = bdexStreamOut(out, X2, VERSION);
             ASSERT(&out == &rvOut2);
             ASSERT(out);
-            const int         LOD2 = out.length();
+            const int         LOD2 = static_cast<int>(out.length());
 
             Out& rvOut3 = bdexStreamOut(out, X3, VERSION);
             ASSERT(&out == &rvOut3);
             ASSERT(out);
-            const int         LOD3 = out.length();
+            const int         LOD3 = static_cast<int>(out.length());
             const char *const OD3  = out.data();
 
             for (int i = 0; i < LOD3; ++i) {
@@ -635,7 +635,7 @@ if (veryVerbose)
             Out out(VERSION_SELECTOR, &allocator);
             out.putInt8(SERIAL_Y);  // Stream out "new" value.
             const char *const OD  = out.data();
-            const int         LOD = out.length();
+            const int         LOD = static_cast<int>(out.length());
 
             Enum mT(X);  const Enum& T = mT;
             ASSERT(X == T);
@@ -661,7 +661,7 @@ if (veryVerbose)
             out.putInt8(SERIAL_Y);  // Stream out "new" value.
 
             const char *const OD  = out.data();
-            const int         LOD = out.length();
+            const int         LOD = static_cast<int>(out.length());
 
             Enum mT(X);  const Enum& T = mT;
             ASSERT(X == T);
@@ -684,7 +684,7 @@ if (veryVerbose)
             out.putInt8(SERIAL_Y);  // Stream out "new" value.
 
             const char *const OD  = out.data();
-            const int         LOD = out.length();
+            const int         LOD = static_cast<int>(out.length());
 
             Enum mT(X);  const Enum& T = mT;
             ASSERT(X == T);
@@ -711,7 +711,7 @@ if (veryVerbose)
             out.putInt8(static_cast<char>(Obj::e_JAN - 1));
 
             const char *const OD  = out.data();
-            const int         LOD = out.length();
+            const int         LOD = static_cast<int>(out.length());
 
             Enum mT(X);  const Enum& T = mT;
             ASSERT(X == T);
@@ -738,7 +738,7 @@ if (veryVerbose)
             out.putInt8(static_cast<char>(Obj::e_DEC + 1));
 
             const char *const OD  = out.data();
-            const int         LOD = out.length();
+            const int         LOD = static_cast<int>(out.length());
 
             Enum mT(X);  const Enum& T = mT;
             ASSERT(X == T);
